requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
langchain==0.1.17
langchain-openai==0.1.1
langchain-pinecone==0.0.4
pinecone-client==3.0.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.2 